+++ src/test/typopo.test.js
@@ -71,7 +71,7 @@ describe('Tests that all modules are plugged in', () => {
 		"Company (p)2017": "Company ℗ 2017",
 
 		// abbreviations
-		"E. g.something": "e.g. something",
+		"E. g.something": "e.g. something",
 
 	};